+++ src/experimental/procedural-generation/src/const/galaxy.ts
@@ -1,0 +1,6 @@
+export const GALAXY = {
+  width: 1800,
+  height: 1200,
+  sectorSize: 16,
+  planetColors: [0xffffff, 0xfcba03, 0x35fc03, 0x05f2ee, 0x9d02eb, 0xed0767]
+};