+++ src/experimental/procedural-generation/src/interfaces/interfaces.ts
@@ -1,0 +1,5 @@
+export interface IStarSystemConstructor {
+  scene: Phaser.Scene;
+  x: number;
+  y: number;
+}